+++ resources/views/admin/space/content/contentlist.blade.php
@@ -21,7 +21,8 @@ foreach ($contentlist as $key => $value) {
                 <thead>
                     <tr>
                         <th style="width:5%"></th>
-                        <th style="width:75%">{{ trans('template_contentlist.title') }}</th>
+                        <th style="width:5%"></th>
+                        <th style="width:70%">{{ trans('template_contentlist.title') }}</th>
                         <th style="width:20%">{{ trans('template_contentlist.date') }}</th>
                     </tr>
                 </thead>
@@ -35,6 +36,11 @@ foreach ($contentlist as $key => $value) {
                           <input type="hidden" name="id" class="id" value="{{ $content['id'] }}">
                           <input type="hidden" name="weight" class="weight" value="{{ $content['weight'] }}">
                         </td>
+                        <td class="field-preview-image">
+												@if (isset($content['preview_image_uri']))
+												<img src="{{ $content['preview_image_uri'] }}" width="40">
+												@endif
+												</td>
                         <td class="field-title">
                             <div style="font-weight:bold;word-wrap:break-word;" class="title">{{ $content['title'] }}</div>
                             <div>